Add support for quality scores in compact reads format. Changed fasta-to-compact and compact-to-fasta to read and write with the Sanger or Illumina quality encoding. Modified aligners to indicate which format they require (bwa needs fastq format, lastag fasta format, lastal fastq format). This will need extensive testing as some of these changes can affect gobyweb.
We don't yet support the Solexa quality score encoding (it is a bit obsolete anyway). Note that Lastal indicates that they support this older format in their documentation. We use the FASTQ-SANGER encoding to communicate with lastal.

http://en.wikipedia.org/wiki/FASTQ_format
http://maq.sourceforge.net/fastq.shtml
http://last.cbrc.jp/last/doc/last-manual.txt (look for FASTQ-SANGER)

Please note that the output format in compact-to-fasta now defaults to Fasta format. This format has no quality scores, and consequently, we now never write quality scores when Fasta is requested. The aligners that need quality scores must request FastaQ format explicitely.

git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@22323 60cd2ccc-2b05-0410-9566-da273a940bff