The case of an insertion and mutation in the same read is more common than expected. We encounter it in the goby sample data.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 39M1I4M), the mdAttribute (42A0) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.

This caused an exception in the code downstream because the reference sequence had length zero. Now fixed. More examples of the insertion match case below.

WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 39M1I4M), the mdAttribute (42A0) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 27M1I16M), the mdAttribute (35G7) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 3M1I40M), the mdAttribute (1C41) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 6M1I37M), the mdAttribute (2C1C38) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 35M2I7M), the mdAttribute (38A3) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 40M1I3M), the mdAttribute (26T16) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 13M1I30M), the mdAttribute (19C15C7) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 4M1I39M), the mdAttribute (1T41) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 4M1I39M), the mdAttribute (2G40) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 36M2I6M), the mdAttribute (39C2) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 7M1I36M), the mdAttribute (11A31) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 2M1I41M), the mdAttribute (37G3T1) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 12M2I30M), the mdAttribute (1C40) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 29M1I14M), the mdAttribute (9G33) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.
WARN  SAMToCompactMode     - When a CIGAR string contains an insertion (i.e., 21M1I22M), the mdAttribute (3T18C20) must indicate only matches for the rest of the sequence. Other cases are not supported at this time. These side cases will be ignored in the input.


git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@22505 60cd2ccc-2b05-0410-9566-da273a940bff