Remove DLL export directives from the template definition - we DON'T want to export these otherwise there may be linker problems.
Remove operator definitions outside of the MessageChunks template - we don't need them in both places.

Check the names of the files as well as the current chunk iterator.  Typically we wouldn't need to do this, but in the case where the chunks are empty (which can happen if the file doesn't exist) comparing iterators to an empty vector may produce unpredictable results.


git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@23931 60cd2ccc-2b05-0410-9566-da273a940bff